wishlist_scraper/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
boto3==1.28.65
botocore==1.31.65
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
importlib-metadata==6.8.0
jmespath==1.0.1
logger==1.4
psycopg2-binary==2.9.9
pyee==8.2.2
pyppeteer==1.0.2
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
soupsieve==2.5
tabulate==0.9.0
tqdm==4.66.1
urllib3==1.26.18
websockets==10.4
zipp==3.17.0
